LayoutTests/platform/ios/http/tests/cookies/same-site/popup-cross-site-post-expected.txt


FAIL '127.0.0.1' is not same-site with 'localhost', so samesite cookies are not sent via POST. assert_equals: implicit-strict expected (undefined) undefined but got (string) "1"
